+++ django2pydantic/__init__.py
@@ -2,8 +2,9 @@
 
 import django_stubs_ext
 
+from django2pydantic.infer import InferredField
 from django2pydantic.registry import FieldTypeRegistry
-from django2pydantic.schema import BaseSchema
+from django2pydantic.schema import BaseSchema, SchemaConfig
 from django2pydantic.types import Infer, InferExcept, ModelFields, ModelFieldsCompact
 
 django_stubs_ext.monkeypatch()
@@ -13,7 +14,9 @@ __all__ = [
     "FieldTypeRegistry",
     "Infer",
     "InferExcept",
+    "InferredField",
     "ModelFields",
     "ModelFieldsCompact",
+    "SchemaConfig",
 ]
-__version__ = "0.5.1"
+__version__ = "0.8.0"