+++ django2pydantic/base.py
@@ -33,6 +33,7 @@ from django2pydantic.types import (
     SetType,
     TDjangoModel,
 )
+from django2pydantic.utils import override_type_and_meta
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -130,11 +131,17 @@ def create_pydantic_model(  # noqa: C901, PLR0912, PLR0915, WPS210, WPS231 # NOS
         #             )(validator)
 
         pydantic_field_info: FieldInfo
-        if field_def is Infer:
+        if field_def is Infer or isinstance(field_def, InferExcept):
             type_handler = field_type_registry.get_handler(django_field)
             python_type = type_handler.get_pydantic_type()
             pydantic_field_info = type_handler.get_pydantic_field()
-            pydantic_fields[field_name] = (python_type, pydantic_field_info)
+
+            if isinstance(field_def, InferExcept):
+                python_type, pydantic_field_info = override_type_and_meta(
+                    pydantic_type=python_type,
+                    field_info=pydantic_field_info,
+                    overrides=field_def,
+                )
 
             if type(django_field) in {  # noqa: WPS516
                 ForeignKey,
@@ -149,16 +156,10 @@ def create_pydantic_model(  # noqa: C901, PLR0912, PLR0915, WPS210, WPS231 # NOS
                     mode="wrap",
                 )(BaseMixins.validate_relation)
 
-        elif isinstance(field_def, InferExcept):
-            type_handler = field_type_registry.get_handler(django_field)
-            python_type = type_handler.get_pydantic_type()
-            pydantic_field_info = type_handler.get_pydantic_field()
-
-            # Override the field values
-            for detail_key, detail_value in field_def.args.items():
-                setattr(pydantic_field_info, detail_key, detail_value)
-
-            pydantic_fields[field_name] = (python_type, pydantic_field_info)
+            pydantic_fields[field_name] = (  # pyre-ignore[6]
+                python_type,
+                pydantic_field_info,
+            )
 
         # If the extracted fields is a type[pydantic.BaseModel]:
         elif isinstance(field_def, type) and issubclass(field_def, BaseModel):