+++ noxfile.py
@@ -3,6 +3,8 @@
 See: https://nox.thea.codes/en/stable/config.html
 """
 
+from pathlib import Path
+
 import nox
 
 # Allow proper use of different python versions
@@ -13,7 +15,21 @@ nox.options.default_venv_backend = "uv"
 @nox.parametrize("django_version", ["4.2", "5.0", "5.1", "5.2"])
 def tests(session: nox.Session, django_version: str) -> None:
     """Run the test suite."""
-    session.install(".", f"django=={django_version}", "--group", "tests")
+    # https://nox.thea.codes/en/stable/cookbook.html#using-a-lockfile
+    _ = session.run_install(
+        "uv",
+        "sync",
+        "--no-dev",
+        "--group=tests",
+        f"--python={session.virtualenv.location}",
+        env={"UV_PROJECT_ENVIRONMENT": session.virtualenv.location},
+    )
+    # Override django version from lockfile
+    session.install(f"django=={django_version}")
+
+    # a fix so .ipynb tests imports succeed:
+    session.env["PYTHONPATH"] = str(Path(__file__).parent.resolve())
+
     _ = session.run("pytest", "-vv")